+++ threedi_schematisation_editor/processing/algorithms_conversion.py
@@ -74,7 +74,7 @@ class ImportConnectionNodes(QgsProcessingAlgorithm):
         self.addParameter(target_gpkg)
 
     def processAlgorithm(self, parameters, context, feedback):
-        source_layer = self.parameterAsSource(parameters, self.SOURCE_LAYER, context)
+        source_layer = self.parameterAsVectorLayer(parameters, self.SOURCE_LAYER, context)
         if source_layer is None:
             raise QgsProcessingException(self.invalidSourceError(parameters, self.SOURCE_LAYER))
         import_config_file = self.parameterAsFile(parameters, self.IMPORT_CONFIG, context)
@@ -147,7 +147,7 @@ class ImportCulverts(QgsProcessingAlgorithm):
         self.addParameter(target_gpkg)
 
     def processAlgorithm(self, parameters, context, feedback):
-        source_layer = self.parameterAsSource(parameters, self.SOURCE_LAYER, context)
+        source_layer = self.parameterAsVectorLayer(parameters, self.SOURCE_LAYER, context)
         if source_layer is None:
             raise QgsProcessingException(self.invalidSourceError(parameters, self.SOURCE_LAYER))
         import_config_file = self.parameterAsFile(parameters, self.IMPORT_CONFIG, context)
@@ -226,7 +226,7 @@ class ImportOrifices(QgsProcessingAlgorithm):
         self.addParameter(target_gpkg)
 
     def processAlgorithm(self, parameters, context, feedback):
-        source_layer = self.parameterAsSource(parameters, self.SOURCE_LAYER, context)
+        source_layer = self.parameterAsVectorLayer(parameters, self.SOURCE_LAYER, context)
         if source_layer is None:
             raise QgsProcessingException(self.invalidSourceError(parameters, self.SOURCE_LAYER))
         import_config_file = self.parameterAsFile(parameters, self.IMPORT_CONFIG, context)
@@ -305,7 +305,7 @@ class ImportWeirs(QgsProcessingAlgorithm):
         self.addParameter(target_gpkg)
 
     def processAlgorithm(self, parameters, context, feedback):
-        source_layer = self.parameterAsSource(parameters, self.SOURCE_LAYER, context)
+        source_layer = self.parameterAsVectorLayer(parameters, self.SOURCE_LAYER, context)
         if source_layer is None:
             raise QgsProcessingException(self.invalidSourceError(parameters, self.SOURCE_LAYER))
         import_config_file = self.parameterAsFile(parameters, self.IMPORT_CONFIG, context)
@@ -384,7 +384,7 @@ class ImportPipes(QgsProcessingAlgorithm):
         self.addParameter(target_gpkg)
 
     def processAlgorithm(self, parameters, context, feedback):
-        source_layer = self.parameterAsSource(parameters, self.SOURCE_LAYER, context)
+        source_layer = self.parameterAsVectorLayer(parameters, self.SOURCE_LAYER, context)
         if source_layer is None:
             raise QgsProcessingException(self.invalidSourceError(parameters, self.SOURCE_LAYER))
         import_config_file = self.parameterAsFile(parameters, self.IMPORT_CONFIG, context)